Move OmnitureAnalyticsData into its own file
