+++ common/app/views/support/package.scala
@@ -90,62 +90,6 @@ object ContributorLinks {
   def apply(html: Html, tags: Seq[Tag])(implicit request: RequestHeader): Html = apply(html.body, tags)
 }
 
-object OmnitureAnalyticsData {
-  def apply(page: MetaData, jsSupport: String, path: String)(implicit request: RequestHeader): Html = {
-
-    val data = page.metaData map {
-      case (key, JsString(s)) => key -> s
-      case (key, jValue: JsValue) => key -> Json.stringify(jValue)
-    }
-    val pageCode = data.getOrElse("pageCode", "")
-    val contentType = data.getOrElse("contentType", "")
-    val section = data.getOrElse("section", "")
-    val platform = "frontend"
-    val publication = data.getOrElse("publication", "")
-    val omnitureEvent = data.getOrElse("omnitureEvent", "")
-    val registrationType = data.getOrElse("registrationType", "")
-    val omnitureErrorMessage = data.getOrElse("omnitureErrorMessage", "")
-
-    val isContent = page match {
-      case c: Content => true
-      case _ => false
-    }
-
-    val pageName = page.analyticsName
-    val analyticsData = Map(
-      ("g", path),
-      ("ns", "guardian"),
-      ("pageName", pageName),
-      ("cdp", "2"),
-      ("v7", pageName),
-      ("c3", publication),
-      ("ch", section),
-      ("c9", section),
-      ("c4", data.getOrElse("keywords", "")),
-      ("c6", data.getOrElse("author", "")),
-      ("c8", pageCode),
-      ("v8", pageCode),
-      ("c9", contentType),
-      ("c10", data.getOrElse("tones", "")),
-      ("c11", section),
-      ("c13", data.getOrElse("series", "")),
-      ("c25", data.getOrElse("blogs", "")),
-      ("c14", data("buildNumber")),
-      ("c19", platform),
-      ("v19", platform),
-      ("v67", "nextgenServed"),
-      ("c30", if (isContent) "content" else "non-content"),
-      ("c56", jsSupport),
-      ("event", omnitureEvent),
-      ("v23", registrationType),
-      ("e27", omnitureErrorMessage)
-    )
-
-
-    Html(analyticsData map { case (key, value) => s"$key=${encode(value, "UTF-8")}" } mkString "&")
-  }
-}
-
 object ContentLayout {
   implicit class ContentLayout(content: model.Content) {
